Move reports dir
