+++ api-controllers/ServiceFabrikAdminController.js
@@ -14,7 +14,7 @@ const utils = require('../broker/lib/utils');
 const fabrik = require('../broker/lib/fabrik');
 const bosh = require('../data-access-layer/bosh');
 const ScheduleManager = require('../broker/lib/jobs');
-const BackupReportManager = require('../broker/lib/reports');
+const BackupReportManager = require('../reports');
 const CONST = require('../common/constants');
 const maintenanceManager = require('../broker/lib/maintenance').maintenanceManager;
 const Conflict = errors.Conflict;
